backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Authentication & Security
djangorestframework-simplejwt==5.5.1
django-ratelimit==4.1.0
cryptography==41.0.7
setuptools<81.0.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# WebSockets
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Cache & Task Queue
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0

# File Storage
django-cloudinary-storage==0.3.0
cloudinary==1.36.0
Pillow==10.1.0

# Email & SMS
django-ses==3.5.0
twilio==8.10.0

# Monitoring & Logging
sentry-sdk==1.38.0

# API Documentation
drf-yasg==1.21.7

# Utilities
python-dateutil==2.8.2
pytz==2023.3
django-filter==23.5
bleach==6.1.0
qrcode[pil]==7.4.2
geopy==2.4.1
icalendar==5.0.11

